backend/requirements.txt
# ─────────────────────────────────────────────
#  Agentic-Support-Automator – Backend Dependencies
# ─────────────────────────────────────────────

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# AI / Agent Orchestration
langgraph==0.1.5
langchain==0.2.0
langchain-openai==0.1.7
langchain-community==0.2.0
langgraph-checkpoint-postgres==1.0.2

# Supabase & Database
supabase==2.7.4
psycopg2-binary==2.9.9
psycopg[binary,pool]==3.1.19
sqlalchemy==2.0.30
alembic==1.13.1

# Embeddings (local, free – no API key needed)
sentence-transformers>=2.7.0

# Configuration & Utilities
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
tenacity==8.3.0

# LLM Providers
groq>=1.0.0
openai>=1.30.0

# Rate Limiting
slowapi>=0.1.9

# Email
resend>=2.0.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
